requirements.txt
mcp>=0.9.0
httpx>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.9.0

